requirements.txt
flask
python-engineio
eventlet
python-socketio
numpy
pyDHE
pandas
tensorflow
keras
sympy
pyaes
crypto
flask_cors
pymemcache
category_encoders
